Add PR action to retrieve labels and assert values
This change adds an action that supports the blocking of a PR if label
requirements are not met.

This action is limited to running on `pull_request` or
`pull_request_target` events and takes the labels from the GitHub. These
labels are then checked against a comma-separated required user input
of labels and shows the diff.

An `allOf` optional parameter can be specified with a Truthy value that
will ensure all the specified labels are against the Pull Request.